Remove and add buttons for admins, security in controllers

+++ lib/microblog_web/controllers/post_controller.ex
@@ -6,21 +6,21 @@ defmodule MicroblogWeb.PostController do
   alias Microblog.Response
 
   def index(conn, _params) do
-    posts = Blog.list_posts()
+      posts = Blog.list_posts()
 
-    render(conn, "index.html", posts: Enum.reverse(posts))
-
+      render(conn, "index.html", posts: Enum.reverse(posts))
   end
 
   def new(conn, _params) do
-    if get_session(conn, :user_id) do
-      changeset = Blog.change_post(%Post{})
-      render(conn, "new.html", changeset: changeset)
-    else
+    cur_user = fetch_session(conn, :current_user)
+    if !cur_user do
       conn
       |> put_flash(:error, "You must log in to create a post")
       |> redirect(to: user_path(conn, :index))
     end
+
+    changeset = Blog.change_post(%Post{})
+    render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"post" => post_params}) do
@@ -37,8 +37,8 @@ defmodule MicroblogWeb.PostController do
   end
 
   def show(conn, %{"id" => id}) do
-    post = Blog.get_post!(id)
-    render(conn, "show.html", post: post)
+      post = Blog.get_post!(id)
+      render(conn, "show.html", post: post)
   end
 
   def edit(conn, %{"id" => id}) do
@@ -48,6 +48,13 @@ defmodule MicroblogWeb.PostController do
   end
 
   def update(conn, %{"id" => id, "post" => post_params}) do
+    cur_user = fetch_session(conn, :current_user)
+    if !cur_user do
+      conn
+      |> put_flash(:error, "You must log in to update a post")
+      |> redirect(to: user_path(conn, :index))
+    end
+
     post = Blog.get_post!(id)
 
     case Blog.update_post(post, post_params) do
@@ -61,11 +68,24 @@ defmodule MicroblogWeb.PostController do
   end
 
   def delete(conn, %{"id" => id}) do
-    post = Blog.get_post!(id)
-    {:ok, _post} = Blog.delete_post(post)
+    cur_user = fetch_session(conn, :current_user)
+    if !cur_user do
+      conn
+      |> put_flash(:error, "You must log in to delete a post")
+      |> redirect(to: user_path(conn, :index))
+    end
 
-    conn
-    |> put_flash(:info, "Post deleted successfully.")
-    |> redirect(to: post_path(conn, :index))
+    post = Blog.get_post!(id)
+
+    if post.user_id == id do
+      {:ok, _post} = Blog.delete_post(post)
+      conn
+      |> put_flash(:info, "Post deleted successfully.")
+      |> redirect(to: post_path(conn, :index))
+    else
+      conn
+      |> put_flash(:error, "Cannot delete another user's post")
+      |> redirect(to: user_path(conn, :index))
+    end
   end
 end
